VPSTradingBot/requirements.txt
# Core
aiohttp>=3.9.0
numpy>=1.26.0
pandas>=2.1.0
yfinance>=0.2.30
python-dotenv>=1.0.0
requests>=2.31.0
colorlog>=6.8.0

# API & Server
fastapi>=0.104.0
uvicorn>=0.24.0

# ML & Data Science
scikit-learn>=1.3.0
joblib>=1.3.0
python-dateutil>=2.8.2

# Utilities
# schedule (optional, if needed later)
